Add support for auto-updating components on DOM change
Fix: #562
