+++ packages/js-toolkit/helpers/createApp.ts
@@ -1,8 +1,6 @@
 import type { Base, BaseConstructor, BaseProps } from '../Base/index.js';
 import type { Features } from '../Base/features.js';
-import { getInstances } from '../Base/index.js';
 import { features } from '../Base/features.js';
-import { useMutation } from '../services/index.js';
 import { isBoolean, isObject, isString } from '../utils/index.js';
 
 export type CreateAppOptions = Partial<Features> & {
@@ -41,25 +39,6 @@ export function createApp<S extends BaseConstructor<Base>, T extends BaseProps =
   if (isObject(attributes)) {
     features.set('attributes', attributes);
   }
-
-  // Terminate components whose root element has been removed from the DOM
-  const service = useMutation(document.documentElement, { childList: true, subtree: true });
-  const symbol = Symbol('createApp');
-  service.add(symbol, (props) => {
-    for (const mutation of props.mutations) {
-      if (mutation.type === 'childList') {
-        for (const node of mutation.removedNodes) {
-          if (!node.isConnected) {
-            for (const instance of getInstances()) {
-              if (node === instance.$el || node.contains(instance.$el)) {
-                instance.$terminate();
-              }
-            }
-          }
-        }
-      }
-    }
-  });
 
   async function init() {
     app = new App(root) as S & Base<T>;
